Add check if the file that is being retrieved on a directory changer exists and skip the copy if it does not exist. This is useful in instances where optional files are requested when executing a physics code and not all files are created because the user settings differ than the default behavior.

+++ armi/utils/directoryChangers.py
@@ -187,6 +187,10 @@ class DirectoryChanger:
 
             for fromName, destName in copies:
                 fromPath = os.path.join(initialPath, fromName)
+                if not os.path.exists(fromPath):
+                    runLog.warning(f"{fromPath} does not exist and will not be copied.")
+                    continue
+
                 toPath = os.path.join(destinationPath, destName)
                 runLog.extra("Copying {} to {}".format(fromPath, toPath))
                 shutil.copy(fromPath, toPath)
